+++ Sources/VCore/Helpers/Architectural Pattern Helpers/SwiftUI/Coordinating Navigation Stack (Observable Object)/CoordinatingNavigationStackOO.swift
@@ -1,26 +1,26 @@
 //
-//  CoordinatingNavigationStack.swift
+//  CoordinatingNavigationStackOO.swift
 //  VCore
 //
-//  Created by Vakhtang Kontridze on 25.09.23.
+//  Created by Vakhtang Kontridze on 19.06.22.
 //
 
 import SwiftUI
 
-// MARK: - Coordinating Navigation Stack
+// MARK: - Coordinating Navigation Stack (Observable Object)
 /// `NavigationStack` that manages `NavigationPath` for representing content in the stack.
 ///
-/// `View` embeds `NavigationStackCoordinator` in the environment that can be used to by contents in the stack.
+/// `View` embeds `NavigationStackCoordinatorOO` in the environment that can be used to by contents in the stack.
 /// Can be used to push or pop `View`s programmatically.
 ///
 ///     var body: some View {
-///         CoordinatingNavigationStack(root: {
+///         CoordinatingNavigationStackOO(root: {
 ///             HomeView()
 ///         })
 ///     }
 ///
 ///     struct HomeView: View {
-///         @Environment(\.navigationStackCoordinator) private var navigationStackCoordinator: NavigationStackCoordinator!
+///         @Environment(\.navigationStackCoordinatorOO) private var navigationStackCoordinator: NavigationStackCoordinatorOO!
 ///
 ///         var body: some View {
 ///             Button(
@@ -34,54 +34,54 @@ import SwiftUI
 ///         }
 ///     }
 ///
-@available(iOS 17.0, macOS 14.0, tvOS 17.0, watchOS 10.0, *)
-public struct CoordinatingNavigationStack<Root>: View where Root: View {
+@available(iOS 16.0, macOS 13.0, tvOS 16.0, watchOS 9.0, *)
+public struct CoordinatingNavigationStackOO<Root>: View where Root: View {
     // MARK: Properties
-    @State private var navigationStackCoordinator: NavigationStackCoordinator
-    private let root: (NavigationStackCoordinator) -> Root
-
+    @StateObject private var navigationStackCoordinator: NavigationStackCoordinatorOO
+    private let root: (NavigationStackCoordinatorOO) -> Root
+    
     // MARK: Initializers - NavigationPath
     /// Initializes `CoordinatingNavigationStackOO`.
     public init(
         path navigationPath: @escaping @autoclosure () -> NavigationPath,
-        @ViewBuilder root: @escaping (NavigationStackCoordinator) -> Root
+        @ViewBuilder root: @escaping (NavigationStackCoordinatorOO) -> Root
     ) {
-        self._navigationStackCoordinator = State(wrappedValue: NavigationStackCoordinator(path: navigationPath()))
+        self._navigationStackCoordinator = StateObject(wrappedValue: NavigationStackCoordinatorOO(path: navigationPath()))
         self.root = root
     }
-
+    
     /// Initializes `CoordinatingNavigationStackOO`.
     public init(
         path navigationPath: @escaping @autoclosure () -> NavigationPath,
         @ViewBuilder root: @escaping () -> Root
     ) {
-        self._navigationStackCoordinator = State(wrappedValue: NavigationStackCoordinator(path: navigationPath()))
+        self._navigationStackCoordinator = StateObject(wrappedValue: NavigationStackCoordinatorOO(path: navigationPath()))
         self.root = { _ in root() }
     }
-
+    
     // MARK: Initializers
     /// Initializes `CoordinatingNavigationStackOO`.
     public init(
-        @ViewBuilder root: @escaping (NavigationStackCoordinator) -> Root
+        @ViewBuilder root: @escaping (NavigationStackCoordinatorOO) -> Root
     ) {
-        self._navigationStackCoordinator = State(wrappedValue: NavigationStackCoordinator(path: NavigationPath()))
+        self._navigationStackCoordinator = StateObject(wrappedValue: NavigationStackCoordinatorOO(path: NavigationPath()))
         self.root = root
     }
-
+    
     /// Initializes `CoordinatingNavigationStackOO`.
     public init(
         @ViewBuilder root: @escaping () -> Root
     ) {
-        self._navigationStackCoordinator = State(wrappedValue: NavigationStackCoordinator(path: NavigationPath()))
+        self._navigationStackCoordinator = StateObject(wrappedValue: NavigationStackCoordinatorOO(path: NavigationPath()))
         self.root = { _ in root() }
     }
-
+    
     // MARK: Body
     public var body: some View {
         NavigationStack(
             path: $navigationStackCoordinator.path,
             root: { root(navigationStackCoordinator) }
         )
-        .environment(\.navigationStackCoordinator, navigationStackCoordinator)
+        .navigationStackCoordinatorOO(navigationStackCoordinator)
     }
 }