+++ Sources/VCore/Helpers/Architectural Pattern Helpers/SwiftUI/Coordinating Navigation Stack/NavigationStackCoordinatorEnvironmentValue.swift
@@ -2,13 +2,13 @@
 //  NavigationStackCoordinatorEnvironmentValue.swift
 //  VCore
 //
-//  Created by Vakhtang Kontridze on 07.08.22.
+//  Created by Vakhtang Kontridze on 26.09.23.
 //
 
 import SwiftUI
 
 // MARK: - Navigation Stack Coordinator Extension
-@available(iOS 16.0, macOS 13.0, tvOS 16.0, watchOS 9.0, *)
+@available(iOS 17.0, macOS 14.0, tvOS 17.0, watchOS 10.0, *)
 extension View {
     func navigationStackCoordinator(
         _ navigationStackCoordinator: NavigationStackCoordinator
@@ -19,20 +19,9 @@ extension View {
 }
 
 // MARK: - Navigation Stack Coordinator Environment Value
-@available(iOS 16.0, macOS 13.0, tvOS 16.0, watchOS 9.0, *)
+@available(iOS 17.0, macOS 14.0, tvOS 17.0, watchOS 10.0, *)
 extension EnvironmentValues {
     /// `NavigationStackCoordinator` of the `View` associated with the environment.
-    ///
-    /// Since `NavigationStackCoordinator` is a reference type, `View` updates won't be triggered.
-    ///
-    ///     @Environment(\.navigationStackCoordinator) private var navigationStackCoordinator: NavigationStackCoordinator!
-    ///
-    ///     var body: some View {
-    ///         CoordinatingNavigationStack(root: {
-    ///             ...
-    ///         })
-    ///     }
-    ///
     public var navigationStackCoordinator: NavigationStackCoordinator? {
         get { self[NavigationStackCoordinatorEnvironmentKey.self] }
         set { self[NavigationStackCoordinatorEnvironmentKey.self] = newValue }
@@ -40,7 +29,7 @@ extension EnvironmentValues {
 }
 
 // MARK: - Navigation Stack Coordinator Environment Key
-@available(iOS 16.0, macOS 13.0, tvOS 16.0, watchOS 9.0, *)
+@available(iOS 17.0, macOS 14.0, tvOS 17.0, watchOS 10.0, *)
 private struct NavigationStackCoordinatorEnvironmentKey: EnvironmentKey {
     static var defaultValue: NavigationStackCoordinator? = nil
 }