+++ Sources/VCore/Services and Managers/Device Orientation Observer/DeviceOrientationObserverOO.swift
@@ -1,36 +1,36 @@
 //
-//  DeviceOrientationObserver.swift
+//  DeviceOrientationObserverOO.swift
 //  VCore
 //
-//  Created by Vakhtang Kontridze on 05.08.23.
+//  Created by Vakhtang Kontridze on 26.09.23.
 //
 
 #if os(iOS) || targetEnvironment(macCatalyst)
 
-import UIKit
+import SwiftUI
 import Combine
 
-// MARK: - Device Orientation Observer
+// MARK: - Device Orientation Observer (Observable Object)
 /// Observer that detects changes in device orientation.
-public final class DeviceOrientationObserver: ObservableObject {
+public final class DeviceOrientationObserverOO: ObservableObject {
     // MARK: Properties
     /// The physical orientation of the device.
-    @Published private(set) public var deviceOrientation: UIDeviceOrientation = DeviceOrientationObserver.getDeviceOrientation()
+    @Published private(set) public var deviceOrientation: UIDeviceOrientation = DeviceOrientationObserverOO.getDeviceOrientation()
 
     private var subscription: AnyCancellable?
 
     // MARK: Initializers
-    /// Initializes `DeviceOrientationObserver`.
+    /// Initializes `DeviceOrientationObserverOO`.
     public init() {
-        setUpObserver()
+        addSubscriptions()
     }
 
     deinit {
         subscription?.cancel()
     }
 
-    // MARK: Observers
-    private func setUpObserver() {
+    // MARK: Subscriptions
+    private func addSubscriptions() {
         subscription = NotificationCenter.default
             .publisher(for: UIDevice.orientationDidChangeNotification)
             .compactMap { _ in Self.getDeviceOrientation() }