+++ Sources/VCore/Views/SwiftUI/View Resetting Container/ViewResetter.swift
@@ -2,29 +2,29 @@
 //  ViewResetter.swift
 //  VCore
 //
-//  Created by Vakhtang Kontridze on 13.04.23.
+//  Created by Vakhtang Kontridze on 26.09.23.
 //
 
 import Foundation
 
 // MARK: - View Resetter
-/// ViewModel placed inside the environment of `ViewResettingContainer`
-/// to trigger view resets on demand.
+/// Object inside the environment of `ViewResettingContainer` to trigger view resets on demand.
 ///
 /// For additional info, refer to `ViewResettingContainer`.
-public final class ViewResetter: ObservableObject {
+@available(iOS 17.0, macOS 14.0, tvOS 17.0, watchOS 10.0, *)
+@Observable public final class ViewResetter {
     // MARK: Properties
-    @Published private(set) var value: Int = 0
-    
+    private(set) var value: Int = 0
+
     // MARK: Initializers
     init() {}
-    
+
     // MARK: Trigger
     /// Triggers reset inside `ViewResettingContainer`.
     public func trigger() {
         value += 1
     }
-    
+
     // MARK: Callable Object
     /// Triggers reset inside `ViewResettingContainer`.
     public func callAsFunction() {