+++ Sources/VCore/Views/SwiftUI/View Resetting Container/ViewResetterEnvironmentValue.swift
@@ -2,23 +2,26 @@
 //  ViewResetterEnvironmentValue.swift
 //  VCore
 //
-//  Created by Vakhtang Kontridze on 23.09.23.
+//  Created by Vakhtang Kontridze on 26.09.23.
 //
 
 import SwiftUI
 
+// MARK: - View Resetter Extension
+@available(iOS 17.0, macOS 14.0, tvOS 17.0, watchOS 10.0, *)
+extension View {
+    func viewResetter(
+        _ viewResetter: ViewResetter
+    ) -> some View {
+        self
+            .environment(\.viewResetter, viewResetter)
+    }
+}
+
 // MARK: - View Resetter Environment Value
+@available(iOS 17.0, macOS 14.0, tvOS 17.0, watchOS 10.0, *)
 extension EnvironmentValues {
-    /// `NavigationStackCoordinatorEnvironmentKey` of the `View` associated with the environment.
-    ///
-    /// Since `ViewResetter` is a reference type, `View` updates won't be triggered.
-    ///
-    ///     @Environment(\.viewResetter) private var viewResetter: ViewResetter!
-    ///
-    ///     var body: some View {
-    ///         ...
-    ///     }
-    ///
+    /// `ViewResetter` of the `View` associated with the environment.
     public var viewResetter: ViewResetter? {
         get { self[ViewResetterEnvironmentKey.self] }
         set { self[ViewResetterEnvironmentKey.self] = newValue }
@@ -26,6 +29,7 @@ extension EnvironmentValues {
 }
 
 // MARK: - View Resetter Environment Key
+@available(iOS 17.0, macOS 14.0, tvOS 17.0, watchOS 10.0, *)
 private struct ViewResetterEnvironmentKey: EnvironmentKey {
     static var defaultValue: ViewResetter? = nil
 }